CLEAN ARCHITECTURE: Rename backend/api to backend/express-handlers
🎯 ELIMINATED CONFUSION: No more duplicate 'api' directories!

✅ RENAMED: backend/api → backend/express-handlers
  - Makes purpose crystal clear (Express.js handlers for local dev)
  - No more confusing duplicate directory names
  - Same comprehensive documentation preserved

✅ UPDATED: All import paths throughout codebase
  - /api re-exports now import from ../backend/express-handlers/
  - Dev server imports from ../express-handlers/
  - All functionality preserved

✅ FINAL CLEAN STRUCTURE:
  /api/                          ← Vercel serverless functions (required by Vercel)
  ├── users/index.js            ← 686 lines (real implementation)
  ├── playlists.js              ← 9-line re-export
  ├── library.js                ← 9-line re-export
  ├── friends.js                ← 9-line re-export
  └── search-users.js           ← 9-line re-export

  /backend/express-handlers/     ← Express.js handlers (clear purpose!)
  ├── users/index.js            ← 686 lines (local dev version)
  ├── playlists.js              ← 946 lines of beautiful documentation ✨
  ├── library.js                ← 503 lines of comprehensive logic ✨
  ├── friends.js                ← 916 lines of complete system ✨
  └── search-users.js           ← 451 lines of detailed functionality ✨

RESULT: Crystal clear codebase with no naming confusion! 🎉
