+++ api/search-users.js
@@ -6,4 +6,4 @@
  */
 
 // Re-export the comprehensive backend implementation
-export { default } from '../backend/api/search-users.js';
+export { default } from '../backend/express-handlers/search-users.js';
